[FIX] web_tour, web_editor: observe iframe DOM mutations in tours
Before this commit, iframes DOM mutations were not taken into account to
update the tour manager. So in the context of mass_mailing, displaying
the website editor in an iframe, the tours were updated at the right
moment (no mutations for 750ms in the global document and in the iframe
document) by chance.

This commit adds an o_iframe class, used to define an iframe that will
render a part of Odoo.
An observer is added in the tour service to monitor the addition of such
iframe (or an iframe container), and sets an observer on that iframe to
correctly update the tour manager, counting the iframe DOM mutations for
the 750ms delay.
In addition to this observer, the iframe's HTML content is added to the
final document HTML log when a tour crashes.

Also, when the tour_service is initialized in an o_iframe, it will not run
the tours.

This fix targets the saas-15.3 as it is needed to avoid a race condition
introduced with [1].

[1]: https://github.com/odoo/odoo/commit/52d32e26c34bef10823582ef5c1f840492b73d7a