[FIX] web_tour, *: observe iframe DOM mutations
*: website, mass_mailing

Before this commit, iframes DOM mutations were not taken into account to
update the tour manager. So in the context of mass_mailing, displaying
the website editor in an iframe, the tours were updated at the right
moment (no mutations for 750ms in the global document and in the iframe
document) by chance.

This commit adds an observer in the tour service to monitor the addition
of an iframe container, and sets an observer on that iframe to correctly
update the tour manager, counting the iframe DOM mutations for the 750ms
delay.

The modules using an iframe can now define their iframe containers by
patching the tour service.

Also, when the tour_service is initialized in an iframe, it will not run
the tours.

task-2687506