[FIX] web,website: fix variable-prefix for website snippet
In BS5 there is a new prefix `bs-` for the root CSS variables.
As some website snippet relay on these values we need to restore the old
behaviour by setting this prefix to an empty string.

[FIX] web,website: Restore negative margins disabled in BS5 by default

[FIX] web: restore BS4 behaviour for setting element content

Fix some cases when we want to render more than one node on the root
level.

In BS4 the jQuery was using all node but in the jQuery compatibility
layer in BS5 it takes only the first node if the jQuery Element is used
in the function. This hack try to restore the old behaviour wrapping all
element in a DIV element (to be single root node).

PS: this can break some querySelector and some CSS selector if they are
too fragile.

e.g. of element that break before this hack:
Popover content for PaymentPopOver XML in ShowPaymentLineWidget

[FIX] web: restore old value from BS4 and adapt to new value

$table-cell-padding -> $table-cell-padding-x
$table-cell-padding -> $table-cell-padding-y

$table-cell-padding-sm -> $table-cell-padding-x-sm
$table-cell-padding-sm -> $table-cell-padding-y-sm

[FIX] web: restore BS4 smooth-scroll behaviour

[FIX] web: restore `text-X` from BS4 that use text-emphasis-variant

In BS5, to generate the text utilities `test-X` its use BS5 utilities
(See $utilities SCSS variables). To restore the old behaviour we replace
with the custom `text-emphasis-variant` mixin.

See odoo/odoo@e111cfd35eb0f0f2fac828cbdafd3b8093d70a1b

[FIX] web: adapt border and background of table

In `_reboot.scss`, BS5 add `border-style` to all element of a table.
In some case we customize the table with CSS rules (e.g. `listview`),
these rules set a size of border. The mix of all rules made that the
border of the `td` in `tfoot` are visible and make a black line.

[FIX] website: reset block quote Y margin to 0

BS5 add a spacer and we reset that by puttin margin-y 0

[FIX] portal: fix all style for table in BS5

In BT5, border color is linked to the text color in it. We doesn't
want that and we restore the same behavior as un BS4 by putting
border-color in grey.

[FIX] website,website_event_exhibitor: restore HR style like in BS4

Since in BS5, the height matter we remove the padding and added into the
margin.

Also, we restore the color of BS4 for the HR tag

> <hr> elements now use height instead of border to better support the
> size attribute. This also enables use of padding utilities to create
> thicker dividers (e.g., <hr class="py-1">).

https://getbootstrap.com/docs/5.1/migration/#content-reboot-etc

[FIX] portal,web,website: restore old default color of BS4

[FIX] website: adjust table layout and color

[FIX] website: restore old value for grid gutter to avoid calc error

In BS5 `$grid-gutter-width` is set in rem and in BS4 it was in px.
So SCSS compiler generates error when it tries to calc rem and px.

[FIX] website: fix breadcrumb padding

restore the default values from BS4

badge restore the padding like in BS4