[FIX] website: fix opening configurator after installing website module
Before that commit, opening the client action was not possible directly
after installing the website module (the client action javscript class
was not yet added to the actions registry).

An action_id param is added to the reload action, so that it can open
the website configurator after reloading the webclient.

task-2687506