[IMP] mail, mass_mailing: remove the <mail.mail> in batch
Purpose
=======
Improve the performance of mass mailing by removing the <mail.mail>, in
batch, all at once. The unlink operation is very expensive for the ORM,
in terms of SQL queries because it needs to check for foreign keys.

We decided to make a table <mail.mail.deletion> to store the mails we
need to delete. We firstly added a new field on the mail, but the
search on this table can be very slow because on the number of record
(and the state field not indexed).

In comparison to a "flag field" on <mail.mail>, this new table cost one
1 SQL query per mail to delete which is already a huge improvement. 

Task-2587345