[FIX] website, *: redirect to the client action when needed
*: website_blog, website_hr_recruitment, website_link, website_sale,
website_slides

This commit changes the redirections from the controllers, that were
always redirecting to the frontend before.

Now, we want to open the website in its backend client action most of
the time (for example, when clicking on the "go to website" buttons on
the form views).

For that, a new get_client_action_url is introduced on the website
model.

task-2687506