[IMP] knowledge: improve permission panel performance
This commit improves the permission panel performance by allowing the method
that retrieves members permissions to receive additional fields to fetch.

The underlying implementation uses a CTE request and joins on the additional
tables to fetch the field/alias pairs given to the method.

This allows simplifying the controller that prepares the permission panel data
by doing a single method call (and a single request).

This should not impact other flows as this "additional_fields" parameter is
optional and has no effect on the quest and the result when not passed.

Task-2674460