[IMP] mass_mailing_sms: add demo data records of subscription contacts
This commit will add demo data of mailing.contact.subscription model.
We need to create seperate records of it rather than directly mapping list_ids
in mailing.contact record because when it directly links in rel table
we are not able to get create_date. So, to get create_date we nee to
create seperate records of them.

TaskId:2702607