[WIP] base: use matching semantics for orm search and filtered_domain
Inconsistencies exist between the orm search and the filtered_domain methods
when including one2many or many2many leafs in the domain. These occur
generally in one of the following situations:

1) When using the logical not on a one2many or a many2many leaf. In general
filtered_domain will adhere to commonly accepted logical not semantics (any
record not matched by the original leaf will match the inverted one) while
the orm search does not. This happens because the orm transforms inverted leafs
by negating the operator inside the leaf. This does not generally work when
multiple values can match the condition.
2) Using negative operators. A similar inconsistency is caused by the logic
used by the filtered_domain method. Negative operators such as != and not in
are implemented by inverting the results of their positive counterparts (any
record is matched by the filtered_domain method if it is not matched by the
corresponding positive operator). Approximately, the positive operators can be
used for matching any related record, while for the negative ones all related
records have to meet the condition.

Additionally, when both situations occur together for the same leaf, the
inconsistency is lifted. However, semantically the results could still be
considered wrong.