[WIP] purchase_(mrp,stock): SVL for price diff on kits
As a SVL is created by component for the kit products, we have to handle
the creation of the SVL for the difference between purchase price and
invoice price.
It will create a SVL for the price diff by component, splitting the diff
between each component depending of their initial price.
For example, at the receipt of a kit of 2 components (Comp A costing 5€
and Comp B costing 15€), it will create two SVL: one for the Comp A and
one for the Comp B.
Then, if we invoice the kit at 60€, it will see we have a diff of 40€.
Again, it will create two SVL and will split the diff between them:
  - 10€ linked to the SVL of the Comp A (25% of the diff);
  - 30€ linked to the SVL of the Comp B (75% of the diff).

To do that, the method `_get_price_difference_svl_values` separates
invoice lines for kits, and this method was moved from `account.move` to
`account.move.line` to filter them efficiently.