[WIP] base: implement recursive search domains
Inconsistencies exist between the orm search and the filtered_domain methods
when including one2many or many2many leafs in the domain. These occur
generally in one of the following situations:

1) When using the logical not on a one2many or a many2many leaf. In general
filtered_domain will adhere to commonly accepted logical not semantics (any
record not matched by the original leaf will match the inverted one) while
the orm search does not. This happens because the orm transforms inverted leafs
by negating the operator inside the leaf. This does not generally work when
multiple values can match the condition.

2) Using negative operators. A similar inconsistency is caused by the logic
used by the filtered_domain method. Negative operators such as != and not in
are implemented by inverting the results of their positive counterparts (any
record is matched by the filtered_domain method if it is not matched by the
corresponding positive operator). Approximately, the positive operators can be
used for matching any related record, while for the negative ones all related
records have to meet the condition.

 Additionally, when both these situations occur together for the same leaf, the
inconsistency is lifted. However, semantically the results could still be
considered wrong. Using the equals operator as an example, the current search
semantics can be summarized as follows:

 For the orm search method:
 * [('x2many_ids.field', '=', value)] -> any =
 * [('x2many_ids.field', '!=', value)] -> any !=
 * ['!', ('x2many_ids.field', '=', value)] -> any !=
 * ['!', ('x2many_ids.field', '!=', value)] -> any =

 For the python side filtered_domain method:
 * [('x2many_ids.field', '=', value)] -> any =
 * [('x2many_ids.field', '!=', value)] -> all !=
 * ['!', ('x2many_ids.field', '=', value)] -> all !=
 * ['!', ('x2many_ids.field', '!=', value)] -> any =

 Where any and all follow their traditional (python) semantics.

3) A less severe inconsistency occurs when specifying conditions on one2many or
many2many fields that use a comodel domain as part of their definition. When
using search this domain is ignored while it is taken into consideration when
using filtered_domain.

This commit aims to provide an answer to these inconsistencies while also
expanding on the expressivity of the domains. As of this commit, leafs are
accepted with an any or all operator and a comodel domain as the right hand
side:

* ['!', ('x2many_ids', 'any', [('field', '=', value)])]
* ['!', ('x2many_ids', 'all', [('field', '=', value)])]

This allows us to express any or all semantics for any operator while also
providing us with the possibility to combine multiple conditions on one2many or
many2many records.