[FIX] web: use the right timezone in patchDate
Before this commit, the date applied by patchDate was using the
machine's timeZone. The tests then depend on the computer on which they
are executed.
In this commit, we will apply the offset stored in Luxon to use the
offset defined in the tests by patchTimeZone.
This fix has no effect on legacy tests.