[FIX] website_hr_recruitment: allow portal users to filter by department
Currently, only a public user or an internal user can access HR
department information. This is causing an access rights error when
trying to filter job offers on the website if connected with a portal
user, while there is no error if not connected (public user). We should
have the same behavior for the two cases.

X-original-commit: 689a3aa9bf08bf3567c0fa0cc4cf67cc3d7e0b54