[FIX] mass_mailing: show values of s_hr in sidebar even if set globally
Local styles (set directly on elements) should always override global
styles set via the Design tab (on #design-element). Therefore, for these
global styles to be able to change things on separator snippets, we need
to remove the width and alignment that were set directly on the elements
themselves by default. This has the unfortunate effect of making it
impossible for the Style tab to know which values to show for width and
alignment (since it finds nothing on the element).
This allows it to figure out the current width and alignment of the s_hr
snippet via the DOM. If no alignment exists, it will default on whatever
the text direction is.