[FIX] web: remove inconstancy and useless rules
In the FieldBinaryImage and FieldBinarySignature widget there is the
`border` attribute to force the image to have a border.
But in the actual environment it's useless as Bootstrap 4 remove de
border of all image by default, so the rules is useless.
But in BS5 this rules was removed so in some place some image as the
border set.

This commit, remove the `border` attribute to restore the BS4 behaviour.

In BS4:
```css
img {
    border-style: none;
}
```

e.g. where there is the case:
Helpdesk > Reporting > Customer Ratings (Kanban card)