[FIX] web_editor, website: prevent race condition when closing s_popup
When clicking on save while having a popup open, in some conditions,
a traceback can appear due to a race condtion.

This is caused by an event listener which updates the visibility status
of said popup in the editor panel. When saving the popup is closed,
and the cleanForSave starts, but while the popup is closing an event
is tirggered which recreates editors. (_onSnippetOptionVisibilityUpdate)

This commit aims at mitigating the appearance of this race condition
by listening to the .hide event instead of the .hidden, whic means
the event is fired right away instead of 400ms later.