[FIX] account: remove invisible on default accounts
Before this commit the "Default accounts" part of settings was only shown if the company had a chart_template_id.
But not having a chart_template_id dooes not mean there is no custom chart of account configured on the company. In that case one must also be able to set default accounts.