[FIX] knowledge: make restricted articles brake the heritage chain
Description of the issue:
=========================

Suppose we have 3 articles with the following structure:

root
   L child
         L grand child

When granting the "write" permission to a user in "root", this user
will inherit this permission in "child" and in "grand child".
When we then restrict this user with the "no access" permission in
child, this permission is inherited by grand-child (correct).
However, the "internal member" permission in "grand child" is still
inherited from "root" (incorrect: should be inherited from "child" as
well, as "child" is now desynchronized from "root").
Then, when a member is invited in the root article, the right is not
herited by "child" nor "grand child" (correct). However, this
permission is shown as being inherited from "root" in the permission
panel in "grand child" (incorrect: "grand child" does not have this
permission, and is in fact inherited from "child" as expected).

Solution:
=========
1. The "inherited_permission_parent_id" and "inherited_permission"
    fields now have "store" set to "False", so that the "internal
    member" permission is correctly updated when restricting a user
    in an article in which his permission was inherited from a parent
    article.

2. Add the boolean "article_is_desynchronized" check in the inner
    join condition of the recursive SQL request used to fetch the
    permissions of an article and its parents.
    This boolean check allows to prune the recursion when an article
    is desynchronized, so that the permissions of the parents of this
    desynchronized article are not fetched, since they are not
    inherited in the desynchronized articles nor in its children.

Task-2866950