[FIX] core: allow usage of newest PyPDF2
The next Debian release will ship PyPDF2 version 2.0.0, this means that
it's already in SID [0].  As there is a work in progress to package Odoo
15.0 for Debian and that it needs to go through SID to be uploaded, we
need to fix Odoo 15.0 in order to work with both PyPDF2 versions.

The first fix is about the PyPDF2 monkey patch that was made in
f8978fc235 in order to fix a bug in PyPDF2. As this bug was fixed in
PyPDF2 in 1.28.0 [1], the monkey patch is not needed anymore when using
2.0.0.

The second one is about the import of exceptions that were in `utils`
before 2.0.0.

[0] https://packages.debian.org/sid/python3-pypdf2
[1] py-pdf/PyPDF2@354f8cefc5cfaa317c5a23f9e943b17bd4b9786c