[IMP] knowledge: allow article drop in the shared section
Currently, dropping an article in the shared section is prevented because if the
dropping destination is at the root of the shared section, the access rights
cannot be inferred.

This commit allows to move an article in the shared section, as long as it is
under another already existing shared article.

It also allows to change the position of a root article within the shared
section.

Replace "is_private" parameter in knowledge.article.move_to method by
"category", since the previous signature did not allow to identify whether an
article was moved as the first article of the workspace section or the first
article of the shared section.

In move_to, infer the category from parent_id or before_article_id or the
provided category in that order to avoid inconsistent requests. Prevent moving
an article as a root of the shared section if it was not already one. Those
checks increase the query count of
KnowledgePerformanceCase.test_article_move_to.

Task-2867052