[IMP] survey: customize preview link design
Purpose:
--------
When sharing URLs on social medias, a link preview is usually shown.
When sharing a survey (or a survey's results), this preview now shows info
related to the survey (survey's title, description, and website), as well as
an image generated on the fly using the survey's information (title, user
and date, background image, some statistics, and "powered by odoo" logo).

The preview is not shown for surveys requiring users to log in, in order
to not disclose info about the survey.

Task-2842682