[FIX] web: auto-save user changes when triggering act.window action
When the user edits a field html, the user can lose its changes by
executing a new `act.window` action within the same tab. This commit
will make sure that the edited record will be properly saved in db
when executing a new `act.window`.

When the user executes a new `act.window`, the system calls the
`canBeRemoved` function of the JS controller via the `beforeLeave` event.
This function will then save the record iif it is marked as "dirty".
To ensure that the record will be marked as dirty and eventually saved
in db, we will ask the rendrer to commit its changes before saving.