[IMP] knowledge: clean up the permission panel
This commit aims to improve the code quality and fix some issues with
the permission panel of Knowledge.

- When the internal permission of an article changes or the permission of
the user changes, the article can move from one category to another. To
reload the tree, we previously reload the whole view which was a quite
expensive. To reduce the loading time, we will now only reload the
permission panel.
- When the user moves an article with a drag-and-drop or uses the "move-to"
modal, the access rights of the article can change. We will therefore
reload the permission panel when the user moves an article.
- When the user is about the "write" permission on the article, we will
save the document before applying the permission changes. That way, the
user will no longer lose their changes.
- The links to the "based on" articles of the permission panel was not
working as the name of the action opening the article was incorrect.
- The loader icon of the permission panel will now be displayed while
the permission panel is loading.

task-2852905