[IMP] knowledge: do not close the share panel when opening the invitation wizard
When the user clicks on the "invite" button of the share panel and clicks
somewhere on the invitation wizard, the permission panel was closing. To
see the new members and their permissions, the user had to re-open the
share panel which was not convenient. To improve the UX experience, we
will keep the permission panel open while the invite wizard is opened.

task-2852905