[IMP] mail: quote only incoming emails content
Purpose
=======
When you reply to an email, the parent email is automatically quoted,
so we are able to hide the parent email under a "Load More" button.

The email quote  has been removed from the HTML sanitizer and is
only done on incoming emails. This is useless in most cases for the HTML
sanitization.

Task-2812488