[FIX] web: select item in the dropdown of statusbar
Before this commit, it is impossible to select an stage folded in
the statusbar component for a task when the user is in the form view.
The reason is the onSelected method has no event passed and so it is
impossible to read `detail` on the event passed.

This commit fixes the issue by using the item variable contained in the
template of statusbar component instead of using an event passed to
onSelected method.