[FIX] web: kanban column creation layout
This commit is a follow-up of odoo/odoo@b72598150501c19af089e79a9557df2b0cda63ad .

It avoids the quick create column to shrink below its base width on
tablet-like screens.

Also, it fixes the scroll of the example columns on mobile-like screens.
By doing so, it allows to scroll to the "example column" but no further,
giving a hint of what a kanban column can look like but not cluttering
too much the view to not "lose" the user in the examples on mobile (as
in this context, the view is column centric and not global like on
desktop).

Part of the SCSS revamp task-2704984

task-2883057