[FIX] mail: Make reply_to take the rendered value of email_from by default for mass mails
When sending a mass mail through the composer, if the field ``reply_to``
had to fall back to ``email_from``, reply_to
would take the value of the template syntax.

This is notably the case when mass-mailing incoices through the accounting app.
Resulting in reply_to fields such as {{user.email}}