[FIX] website: fix page properties dialog
The goal of this commit is to reimplement the same behaviour from
master on page properties to warn user if he created a 'website.page'
with supported mimetype and changes its name (page key will be updated
too).

Since we can edit pages from a website different to the current one
(e.g. in backend page manager), we need to get the 'page.website_id'
of a specific page (e.g. to check where the page is t-called in
'page_search_key_dependencies').

task-2687506