[FIX] website_sale: add right filter for included taxes option
Steps to reproduce:
- Install ecommerce
- Set a tax on the product with the highest price
- Add the "filter by price" customize option on the shop
- Set the filter buttons to the highest prize

Current behavior:
-The highest price of the slider is the one of the product without
taxes and the product is displayed with a price higher than the
one of the slider.

Expected behavior:
-The highest price of the slider is the one of the product with
taxes and the product is displayed with the same price as the
one of the slider.

Explanation:
So far the code only handled the slider with tax excluded option,
they used sql domains to filter with the right taxes. For the tax
included it can not be done that way as the computation of the tax
value can not be done in sql so we have to use some (less efficient)
python logic.

opw-2877126