[IMP] base: add a new "Properties" field
Purpose
=======
Add a new field, "Properties" to be able to light customization of
work flows based on a parent model (e.g. add new properties on tasks
of a specific project).

Specification
=============
The field has some attributes
- `parent_field` which define the field on the current model that
   define the parent
- `parent_properties_field`, the `PropertiesDefinition` field on the
   parent model that contains the properties definition (types, default
   value, model for relational properties,...)

Technical
=========

Parent | Properties definition
------------------------------
The properties definition is stored on the parent, on a JSON field.
This definition contains the type of the properties, the default value,
the model of the many2one,...
```
[
    {
        'name': 'name',
        'string': 'Name',
        'type': 'char',
        'default': 'Default Name',
    }, {
        'name': 'partner_id',
        'string': 'Partner',
        'type': 'many2one',
        'comodel': 'res.partner',
    },
]
```

Child | Properties values
-------------------------
The value is stored on the child, using a Properties field.
```
{
    'name': 'Mitchel',
    'partner_id': 1337,
}
```

When we read this field, we will automatically read the definition on
the parent, and merge both JSON into one, so the web client has the
value of each property, and their definition.

```
[
    {
        'name': 'name',
        'string': 'Name',
        'type': 'char',
        'default': 'Default Name',
        'value': 'Mitchel',
    }, {
        'name': 'partner_id',
        'string': 'Partner',
        'type': 'many2one',
        'comodel': 'res.partner',
        'value': 1337,
    },
]
```

Integrity
---------
If we remove a property on the parent, we won't update the child value.

Instead, when we read the child properties, we will filter them based
on the parent. So the removed properties will be removed the next time
we write on the field.

In the same logic, the many2one existence is checked when we read the
field. There's no foreign key between the integer stored in the JSON
in the SQL row corresponding to the record in database.

Write
-----
We can write on the Properties field with a list of field definition
+ value.

Some types are not JSONifiable (like the date, datetime), they are
stored as string in database and parsed when we read the value.

In order to update the parent definition by writing on the child,
you need to add the dict key `definition_changed` or
`definition_deleted`. This is because we need to be able to know
if the definition has been changed without doing extra SQL queries.

Access rights
-------------
A user can add a many2one / many2many property to a model only if he
has the access rights to it.

Many2one
--------
The model choice of a many2one property is a bit tricky...

1/ For the parent, the web client needs the id of the model and the model
   description to be able to display it like a simple many2one (so the
   user can search model by name, etc). It also needs the model name
   ("res.partner") for the many2one of the default value.
2/ For the child, the web client only needs the name of the model
   (e.g. "res.partner")

Because we will more often read / edit the child (and less the parent),
and because we can not store the description of the model (it can be
translated), only the model name ("res.partner") is stored in the
database.

Every time we select a new model on the properties definition (on the
parent), we will need to fetch the model name anyway to update the
many2one of the default value...

When we set a record on a many2one, we need to store the model on the
child as well, because if we change the model on the parent, the ID
might point to a different record, to this new model, instead of
pointing to False. So when we read the many2one on the child, if the
`original_model` on the child is different from the `model` on the
parent, it means that the model has been changed, and so the many2one
value must be set to False. We don't really duplicate the information,
the model on the child correspond to the model when we set the many2one
value, but the model on the parent correspond to the current model in
the properties definition.
```
{
    'name': 'Mitchel',
    'partner_id': (1337, 'res.partner'),
}
```