[WIP] base: add a new "Properties" field
Purpose
=======

Specification
=============

Technical
=========

Parent | Properties definition
------------------------------
The properties definition is stored on the parent, on a JSON field.
This definition contains the type of the properties, the default value,
the model of the many2one,...
```
[
    {
        'id': 'name',
        'string': 'Name',
        'type': 'char',
        'default': 'Default Name',
    }, {
        'id': 'partner_id',
        'string': 'Partner',
        'type': 'many2one',
        'model': 'res.partner',
    },
]
```

Child | Properties values
-------------------------
The values is stored on the child, using a Properties field.
```
{
    'name': 'Mitchel',
    'partner_id': 1337,
}
```

When we read this field, we will automatically read the definition on
the parent, and merge both JSON into one, so the web client has the
value of each property, and their definition.

```
[
    {
        'id': 'name',
        'string': 'Name',
        'type': 'char',
        'default': 'Default Name',
        'value': 'Mitchel',
    }, {
        'id': 'partner_id',
        'string': 'Partner',
        'type': 'many2one',
        'model': 'res.partner',
        'value': 1337,
    },
]
```

Integrity
---------
If we remove a property on the parent, we won't update the child value.

Instead, when we read the child properties, we will filter them based
on the parent. So the removed properties will be removed the next time
we write on the field.

In the same logic, the many2one existence is checked when we read the
field. There's no foreign key between the integer stored in the JSON
the the SQL row corresponding to the record in database.

Search
------
Postgre offers an operator to search on JSON column, `->>`.

We can search in JSON type using the domain syntax
```
[('properties.name', 'ilike', 'Bob')]
```

Obviously, we can not search for "many2one" field in a properties
"properties.partner_id.name" (because they are juste integer...).

Write
-----
We can write on the Properties field with
- a list of field definition + value
- a dict of {property_name: property_value}

Some types are not JSONifiable (like the date, datetime), they are
stored as string in database and parsed when we read the value.