Revert auto fill partner_id
It messes up because it is a shadow field of the related move, and it causes some onchange/depends happen on the move
result in recalculating narration which can't be stopped by removing the field from to_compute (in the write or in partner_id's
compute methods). So I remove it because it is not in the spec.