[FIX] web_editor: allow to reset background color
Before this commit:

We were unable to reset the background color when text is copy paste from
the discord and use toolbar to select background color but click on trash
icon to reset background color.

After this commit:

Now we can reset the background color when text is copy paste from the discord
and use toolbar to select background color and click on trash icon to reset
background color.

Task-2889682