fixup! [FIX] web: view_button_hook can take a localized beforeExecute as params of the function
This rev allows the function built by the useViewButtonHook to take a beforeExecute callback at call time,
in addition to the existing beforeExecuteAction callback that we can pass to the hook, at the function's build time.

The rationale for this is to allow any caller to prevent the execution of the action represented by the clickParams at the call level.
It is usefull to check if a user is allowed business-wise to execute an action.